refactor(audit): f-8 split core/utils.py into focused modules
#115 finding 8 / #113 A5: core/utils.py was a three-domain grab-bag
(matplotlib font setup, WAV I/O + DSP, plot helpers). Split into:
- core/audio_io.py: to_db, db_to_gain, convolve, dB_unweight, read_wav,
  write_wav, magnitude_response (moved VERBATIM — rfft path pinned by
  test_magnitude_response_parity and the BRIR md5), running_mean.
- core/font_setup.py: the Korean-font setup block + the _font_configured /
  font_setup_result lazy state and set_matplotlib_font.
- core/plotting_utils.py: sync_axes, get_ylim, versus_distance,
  optimize_png_size, save_fig_as_png, config_fr_axis, ADAPTIVE_PALETTE.

core/utils.py is now a re-export shim preserving every public name for the
25+ importers (and the FFmpeg re-exports). The set_matplotlib_font memo gate
moved with it, so gui_main's smoke test and test_gui's bundled-font test now
reset core.font_setup._font_configured (the gate's new home) — all assertions
preserved.

Pure code move: parity test green, demo md5 unchanged (07eef9ef…), full
suite 195 pass / 0 fail / 3 skip, ruff clean. CLAUDE.md core map updated.
